Extract vwc_hourly functionality into separate module
